requirements.txt
flask>=2.0.1,<3.0.0
google-generativeai>=0.3.0,<0.4.0
Pillow>=9.5.0,<10.0.0
python-dotenv>=0.19.0,<0.20.0
flask-limiter>=3.3.0,<4.0.0
flask-talisman>=1.0.0,<2.0.0
gunicorn>=20.1.0
flask-wtf>=1.0.0
pytest>=7.0.0 